🚧 wip(main): see changes v 0.21

+++ src/cm_htable/cm_htable.h
@@ -6,7 +6,7 @@
 //   By: rgramati <rgramati@student.42angouleme.fr  +#+  +:+       +#+        //
 //                                                +#+#+#+#+#+   +#+           //
 //   Created: 2024/10/06 19:06:48 by rgramati          #+#    #+#             //
-//   Updated: 2024/11/07 18:41:41 by rgramati         ###   ########.fr       //
+//   Updated: 2024/11/18 18:31:47 by rgramati         ###   ########.fr       //
 //                                                                            //
 // ************************************************************************** //
 
@@ -30,8 +30,8 @@ struct s_cm_htable
 {
 	struct s_cm_entry	*entries;
 	void				(*del)(void *);
-	uint16_t			size;
-	uint16_t			capacity;
+	uint64_t			size;
+	uint64_t			capacity;
 };
 
 // ************************************************************************** //
